♻️🎪 Rename action schema file
