+++ apps/show/src/routes/_exhibitor.exposants.$token._config.stand.modifier-stand/form.ts
@@ -2,7 +2,7 @@ import { createStrictContext } from "@animeaux/core";
 import { useForm as useFormBase } from "@conform-to/react";
 import { getZodConstraint, parseWithZod } from "@conform-to/zod";
 import { useActionData, useLoaderData } from "@remix-run/react";
-import { ActionSchema } from "./action";
+import { ActionSchema } from "./action-schema";
 import type { action } from "./action.server";
 import type { loader } from "./loader.server";
